A working uploader
